+++ braindecode/models/__init__.py
@@ -33,6 +33,7 @@ from .eegnex import EEGNeX
 from .tsinception import TSceptionV1
 from .eegtcnet import EEGTCNet
 from .syncnet import SyncNet
+from .sccnet import SCCNet
 from .fbcnet import FBCNet
 from .ifnet import IFNetV2
 from .eegminer import EEGMiner