+++ docs/whats_new.rst
@@ -55,6 +55,7 @@ Enhancements
 - Adding :class:`braindecode.models.IFNet` (:gh:`660` by `Bruno Aristimunha`_ )
 - Creating the FilterBank module for new models (:gh:`656` by `Bruno Aristimunha`_ )
 - Including PytorchAudio as dependency and remove copied code (:gh:`661` by `Bruno Aristimunha`_)
+- Adding :class:`braindecode.models.SCCNet` (:gh:`668` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.FBCNet` (:gh:`660` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.FBCNet` (:gh:`657` by `Bruno Aristimunha`_ )
 - Adding :class:`braindecode.models.FBMSNet` (:gh:`658` by `Bruno Aristimunha`_ )